Move some logic from stock_update.js to separate stock_db module
