+++ update/stock_update.js
@@ -1,48 +1,9 @@
-//---CONFIGURE AUTHENTICATION HERE--------------------
-var CLOUDANT_KEY      = 'XXXXXXXXX';
-var CLOUDANT_PASSWORD = 'YYYYYYYYY';
-var CLOUDANT_ACCESS   = 'someurl-bluemix.cloudant.com'
-var DB_NAME = 'DB_NAME';
-//----------------------------------------------------
 
-var Cloudant = require('cloudant');
+var stock_db = require('./stock_db.js');
 var utils = require('./utils.js');
-
-var cloudant = Cloudant({
-    account: CLOUDANT_ACCESS,
-    key: CLOUDANT_KEY,
-    password: CLOUDANT_PASSWORD
-});
-var db = cloudant.db.use(DB_NAME);
 
 //TODO these should be the companies' tickers
 var companies = ['A', 'B', 'C', 'D'];
-
-function getDocs(callback) {
-    db.list({
-        //this field is needed to return all doc data
-        include_docs: true
-    }, function(err, data) {
-        if (utils.isFunc(callback)) {
-            var docs = data.rows.map(function(row) {
-                return row.doc;
-            });
-            callback(err, docs);
-        }
-    });
-}
-
-//include _id and _rev of existing doc in the doc to perform an update
-function insertOrUpdateDoc(doc, callback) {
-
-    db.insert(doc, function(err, body, header) {
-        if (err) {
-            console.log('insertion failed: ' + err.message);
-        } else if (utils.isFunc(callback)) {
-            callback(body, header);
-        }
-    });
-}
 
 function findStockDatum(stocks, company) {
 
@@ -74,7 +35,7 @@ function updateStocksData(articleData, stockData) {
         //TODO batch insert?
         console.log('Inserting into company ' + articleDatum.company + ' article: ' );
         console.log(articleDatum.article);
-        insertOrUpdateDoc(stockDatum);
+        stock_db.insertOrUpdateDoc(stockDatum);
     }
 }
 
@@ -103,7 +64,7 @@ function run() {
 
     getArticleDataForCompanies(companies, function(articlesErr, articleData) {
         if (!articlesErr) {
-            getDocs(function(docsErr, stockData) {
+            stock_db.getDocs(function(docsErr, stockData) {
                 if (!docsErr) {
                     updateStocksData(articleData, stockData);
                 } else {
